Create Spotify Top 50 analysis
I analysed data from a dataset found on Kaggle, the Spotify Top 50 songs from 2021. I selected, and aggregated the data to determine which artists had the highest average popularity, and which song keys had the highest average popularity. I also updated the table in order to return the song keys in their traditional letter designations and to return the key mode as "major" or "minor" rather than the number designations that were supplied in the data.